requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
docutils==0.20.1
fonttools==4.42.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jaraco.classes==3.3.0
keyring==24.2.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
numpy==1.25.2
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pkginfo==1.9.6
plotly==5.16.1
Pygments==2.16.1
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
readme-renderer==41.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
scipy==1.11.1
SimpleWebSocketServer==0.1.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==2.0.4
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.6.1
whitenoise==6.5.0
zipp==3.16.2
